refactor: combine checks related to purity
