+++ packages/safe-ds-lang/tests/resources/validation/purity/duplicate impurity reason/no impure annotation.sdstest
@@ -1,4 +1,4 @@
-package tests.validation.builtins.annotations.impure.duplicateReason
+package tests.validation.purity.duplicateImpurityReason
 
 // $TEST$ no warning r"The impurity reason '.*' was set already\."
